Хранение в файлах переведено на БД.
