+++ .gitignore
@@ -38,3 +38,4 @@ README.md
 LICENSE.md
 
 .htaccess
+/config/db_my.php
