--- base/models/DayModel.php
@@ -1,15 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: geras
- * Date: 19.04.2019
- * Time: 20:12
- */
-
-namespace app\base\models;
-
-
-class DayModel
-{
-
-}
