+++ config/web.php
@@ -1,8 +1,11 @@
 <?php
 
 $params = require __DIR__ . '/params.php';
-//$db = require __DIR__ . '/db.php';
-$db = require __DIR__ . '/db_local.php';
+
+if(file_exists(__DIR__ . '/db_my.php'))
+    $db = require __DIR__ . '/db_my.php';
+else
+    $db = require __DIR__ . '/db.php';
 
 $config = [
     'id' => 'basic',
