--- helpers/storage/activities.txt
@@ -1,1 +1,0 @@
-a:2:{s:4:"data";a:2:{i:0;a:9:{s:5:"title";s:23:"Занятие Yii №4";s:11:"description";s:78:"Четветое занятие по изучению framework Yii2. В 19:00";s:9:"startDate";s:10:"2019-04-25";s:7:"endDate";s:10:"2019-04-25";s:5:"email";s:12:"test@test.ru";s:10:"isBlocking";s:1:"0";s:16:"needNotification";s:1:"1";s:12:"uploadedFile";a:2:{i:0;s:32:"/images/yii2_2_5cc0eb5d789d4.jpg";i:1;s:30:"/images/yii2_5cc0eb5d78aee.jpg";}s:2:"id";i:1;}i:1;a:9:{s:5:"title";s:30:"Съездить на дачу";s:11:"description";s:127:"Съездить на дачу, на субботник. Привести участок в порядок после зимы.";s:9:"startDate";s:10:"2019-04-27";s:7:"endDate";s:10:"2019-04-28";s:5:"email";s:0:"";s:10:"isBlocking";s:1:"1";s:16:"needNotification";s:1:"0";s:12:"uploadedFile";a:2:{i:0;s:80:"D:/OpenServerLast/mydomains/local.yii-app.ru/web/images/yii2_2_5cc0f1bf5c5cb.jpg";i:1;s:78:"D:/OpenServerLast/mydomains/local.yii-app.ru/web/images/yii2_5cc0f1bf5c6c0.jpg";}s:2:"id";i:2;}}s:7:"counter";i:2;}
